requirements.txt
langchain
langchain-core
langchain-community # For openrouter api
langchain-chroma
langchain-text-splitters
sentence-transformers # For embedding models
chromadb
gradio
streamlit
python-dotenv
